--- rdf/src/main/scala/AsyncSPARQLEngine.scala
@@ -1,8 +1,0 @@
-package org.w3.banana
-
-import akka.dispatch._
-
-/**
- * TODO
- */
-trait AsyncSPARQLEngine[Rdf <: RDF] extends MSPARQLEngine[Rdf, Future]
